Moved beacon_chain/utils/simpleserialize.py to ssz/ssz.py
