+++ ssz/__init__.py
@@ -1,0 +1,7 @@
+from .ssz import (  # noqa: F401
+    deepcopy,
+    deserialize,
+    eq,
+    serialize,
+    to_dict,
+)
